Add GitHub Connector
Add a connector to GitHub via their HTTP API

Treat repos and branches as parent directories. Syntax looks like /repo/branch/parentFolder/folder
With an empty path it list all repos. A path of one level list all branches in one repo. Further path list files and folder in this repo, on this branch and at this exact path.

This add a constraint on `PUT` command because we can only create a file on the 3rd level. If less are provided, it will return an error.

Close #35